+++ src/Command/Argument/Search/AggregateArguments.php
@@ -97,6 +97,30 @@ class AggregateArguments extends CommonArguments
     }
 
     /**
+     * Adds a COLLECT reducer to the current GROUPBY.
+     *
+     * Emits "REDUCE COLLECT <narg> <tokens...> [AS <alias>]" where <narg> is the
+     * number of tokens produced by the given CollectArguments (the FIELDS,
+     * DISTINCT, SORTBY and LIMIT clauses) and excludes the trailing AS <alias>.
+     *
+     * @param  CollectArguments $arguments COLLECT clause builder
+     * @param  string|null      $alias     Optional reducer output column name
+     * @return $this
+     */
+    public function reduceCollect(CollectArguments $arguments, ?string $alias = null): self
+    {
+        $tokens = $arguments->toArray();
+
+        array_push($this->arguments, 'REDUCE', 'COLLECT', count($tokens), ...$tokens);
+
+        if ($alias !== null) {
+            array_push($this->arguments, 'AS', $alias);
+        }
+
+        return $this;
+    }
+
+    /**
      * Sorts the pipeline up until the point of SORTBY, using a list of properties.
      *
      * @param  int    $max